requirements.txt
python-telegram-bot==13.15
requests==2.32.5
urllib3==2.5.0
python-dotenv==1.0.0
